requirements.txt.txt
flask
google-api-python-client
python-dotenv
requests
pytrends
gunicorn
